Refactor loaders to setup.
