+++ dart/async/test/html_test.dart
@@ -14,32 +14,15 @@
 @TestOn('browser')
 library pageloader.test.html_no_shadow_dom;
 
-import 'dart:async';
-import 'dart:html' show Element;
-
 import 'package:pageloader/html.dart';
 
 import 'package:test/test.dart';
 
 import 'data/html_setup.dart' as html_setup;
-import 'setup/html_test_setup.dart' show runTests, syncFn;
-import 'src/shared.dart' as shared;
+import 'setup/html_test_setup.dart' show HtmlLoader, runTests, syncFn;
 
 void main() {
-  runTests(pageloaderFactory);
-}
-
-PageLoader pageloaderFactory() {
-  var div = html_setup.setUp();
-  return new HtmlPageLoader(div, executeSyncedFn: syncFn);
-}
-
-class HtmlLoader extends shared.Loader {
-  final HtmlPageLoader loader;
-  HtmlLoader(this.loader);
-
-  Future<T> getInstance<T>(Type type, [dynamic context]) =>
-      loader.getInstance(type);
+  runTests(pageLoaderFactory);
 }
 
 HtmlLoader pageLoaderFactory() => new HtmlLoader(
