+++ dart/async/test/setup/webdriver_test_setup.dart
@@ -18,6 +18,7 @@ import 'dart:async';
 import 'dart:io';
 import 'dart:mirrors' show currentMirrorSystem;
 
+import 'package:pageloader/webdriver.dart';
 import 'package:path/path.dart' as path;
 import 'package:test/test.dart';
 import 'package:webdriver/io.dart' show Capabilities, WebDriver, createDriver;
@@ -77,4 +78,12 @@ Future<WebDriver> _createTestDriver() {
   }
 
   return createDriver(desired: capabilities);
+}
+
+class AsyncLoader extends shared.Loader {
+  final WebDriverPageLoader loader;
+  AsyncLoader(WebDriver driver) : loader = new WebDriverPageLoader(driver);
+
+  Future<T> getInstance<T>(Type type, [dynamic context]) =>
+      loader.getInstance(type);
 }
